// src/lib/types/states/states.ts
export type StateType =
  | 'active'
  | 'active_filled'
  | 'active_filled_hide'
  | 'completed'
  | 'default'
  | 'default_selected'
  | 'default_unselected'
  | 'disabled'
  | 'disabled_empty'
  | 'disabled_filled'
  | 'disabled_filled_hide'
  | 'disabled_selected'
  | 'disabled_unselected'
  | 'error'
  | 'error_empty'
  | 'error_filled'
  | 'error_filled_hide'
  | 'empty'
  | 'filled'
  | 'filled_hide'
  | 'focus'
  | 'hover'
  | 'loading'
  | 'multiple_selected'
  | 'multiple_selected_hover'
  | 'prefilled'
  | 'pressed'
  | 'selected'
  | 'selected_hover'
  | 'success'
  | 'warning'
  | 'visited'
  | 'inactive'
  | 'filling'
  | 'blocked_by_system'
  | 'error_unselected'
  | 'error_selected'
  | 'current'
  | 'current_day'
  | 'start_date_range'
  | 'end_date_range'
  | 'midle_date_range'
  | 'pending'
  | 'unselected';

export const STATES = {
  ACTIVE: 'active',
  ACTIVE_FILLED: 'active_filled',
  ACTIVE_FILLED_HIDE: 'active_filled_hide',
  BLOCKED_BY_SYSTEM: 'blocked_by_system',
  COMPLETED: 'completed',
  CURRENT: 'current',
  CURRENT_DAY: 'current_day',
  DEFAULT: 'default',
  DEFAULT_SELECTED: 'default_selected',
  DEFAULT_UNSELECTED: 'default_unselected',
  DISABLED: 'disabled',
  DISABLED_EMPTY: 'disabled_empty',
  DISABLED_FILLED: 'disabled_filled',
  DISABLED_FILLED_HIDE: 'disabled_filled_hide',
  DISABLED_SELECTED: 'disabled_selected',
  DISABLED_UNSELECTED: 'disabled_unselected',
  EMPTY: 'empty',
  END_DATE_RANGE: 'end_date_range',
  ERROR: 'error',
  ERROR_EMPTY: 'error_empty',
  ERROR_FILLED: 'error_filled',
  ERROR_FILLED_HIDE: 'error_filled_hide',
  ERROR_SELECTED: 'error_selected',
  ERROR_UNSELECTED: 'error_unselected',
  FILLED: 'filled',
  FILLED_HIDE: 'filled_hide',
  FILLING: 'filling',
  FOCUS: 'focus',
  HOVER: 'hover',
  INACTIVE: 'inactive',
  LOADING: 'loading',
  MIDLE_DATE_RANGE: 'midle_date_range',
  MULTIPLE_SELECTED: 'multiple_selected',
  MULTIPLE_SELECTED_HOVER: 'multiple_selected_hover',
  PENDING: 'pending',
  PREFILLED: 'prefilled',
  PRESSED: 'pressed',
  SELECTED: 'selected',
  SELECTED_HOVER: 'selected_hover',
  START_DATE_RANGE: 'start_date_range',
  SUCCESS: 'success',
  UNSELECTED: 'unselected',
  VISITED: 'visited',
  WARNING: 'warning',
} as const;